Move PLP actions and parsing to the merlins connector
