+++ web/app/containers/product-list/container.js
@@ -1,5 +1,6 @@
 import React from 'react'
 
+import * as commands from '../../integration-manager/commands'
 import {isRunningInAstro} from '../../utils/astro-integration'
 import ProductListHeader from './partials/product-list-header'
 import ProductListContents from './partials/product-list-contents'
@@ -13,6 +14,10 @@ const ProductList = () => {
             <ProductListContents />
         </div>
     )
+}
+
+ProductList.fetcher = (url, dispatch) => {
+    dispatch(commands.fetchProductListData(url))
 }
 
 export default ProductList
