+++ web/app/integration-manager/responses.js
@@ -5,16 +5,13 @@ import {addToCartComplete} from '../containers/product-details/actions'
 import {PRODUCT_DETAILS_ITEM_ADDED_MODAL} from '../containers/product-details/constants'
 
 export const receivePdpProductData = createAction('Receive PDP product data')
-
 export const receivePdpUIData = createAction('Receive PDP UI data')
-
 export const receiveCheckoutShippingData = createAction('Receive Checkout Shipping data')
-
 export const receiveCheckoutData = createAction('Receive Checkout Data')
-
 export const receiveHomeData = createAction('Receive Home Data')
-
 export const receiveNavigationData = createAction('Receive Navigation Data')
+export const receiveProductListProductData = createAction('Receive ProductList product data')
+export const receiveCategory = createAction('Receive Category Data')
 
 export const onShippingEmailRecognized = () => {
     return (dispatch) => {
