+++ web/app/store/categories/reducer.js
@@ -1,7 +1,7 @@
 import Immutable from 'immutable'
 import {handleActions} from 'redux-actions'
 import {mergePayloadForActions} from '../../utils/reducer-utils'
-import {receiveCategory} from './actions'
+import {receiveCategory} from '../../integration-manager/responses'
 
 const initialState = Immutable.Map()
