--- web/app/store/products/actions.js
@@ -1,11 +1,0 @@
-import {createAction} from '../../utils/utils'
-
-import {productListParser} from './parser'
-
-export const receiveProductListProductData = createAction('Receive ProductList product data')
-// export const receiveProductDetailsProductData = createAction('Receive ProductDetails product data')
-
-export const processProductList = ({payload: {$, $response}}) =>
-    receiveProductListProductData(productListParser($, $response))
-// export const processProductDetails = ({payload: {$, $response, url}}) =>
-    // receiveProductDetailsProductData({[urlToPathKey(url)]: productDetailsParser($, $response)})
