+++ web/app/store/products/reducer.js
@@ -1,8 +1,7 @@
 import Immutable from 'immutable'
 import {handleActions} from 'redux-actions'
 import {mergePayloadForActions} from '../../utils/reducer-utils'
-import {receivePdpProductData} from '../../integration-manager/responses'
-import {receiveProductListProductData} from './actions'
+import {receivePdpProductData, receiveProductListProductData} from '../../integration-manager/responses'
 
 const initialState = Immutable.Map()
